requirements.txt
django==2.2
djangorestframework==3.9.2
requests===2.22.0